// src/app/models/enuns/Categoria.ts
export enum Categoria{
    ESPORTE = 1,
    VETERANO = 2,
}

// No Jd Cruzeiro temos duas Categorias 1=Esporte e 2=Veterano

// porém para referência..

// Sub-7 para atletas de 6 e 7 anos.
// Sub-8 para atletas de 8 anos.
// Sub-9 para atletas de 8 e 9 anos.
// Sub-11 para atletas de 10 e 11 anos.
// Sub-13 para atletas de 12 e 13 anos.
// Sub-15 para atletas de 14 e 15 anos.
// Sub-17 para atletas de 16 e 17 anos.
// Sub-20 para atletas de 18, 19 e 20 anos.
// Adulto para atletas de 20 anos em diante.
// Veterano para atletas a partir dos 35 anos

// São conhecidas também como: 
// fraldinha (7 a 9 anos), 
// dente de leite (10 a 11 anos), 
// pré-mirim (11 a 12 anos), 
// mirim (12 a 13 anos), 
// infantil (14 a 15 anos), 
// infanto-juvenil (15 a 16 anos), 
// juvenil (17 a 18 anos) e júnior (17 a 20 anos).